requirements.txt
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
faker
sentry-sdk
pytest
pytest-asyncio
pytest-cov
httpx
requests
elasticsearch
openai
alembic
python-multipart
pytest-mock
